Make hash method always returns a signed integer

+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/JavaStringHash.java
@@ -20,7 +20,7 @@ package org.apache.pulsar.client.impl;
 
 public class JavaStringHash implements Hash {
     @Override
-    public long makeHash(String s) {
+    public int makeHash(String s) {
         return s.hashCode() & Integer.MAX_VALUE;
     }
 }
